src/main/resources/python/requirements.txt
flask
pdfminer.six
